requirements.txt
Django==5.2.4
python-dotenv==1.1.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.0
django-filter==25.1